pyproject.toml
[project]
name = "reviewly"
version = "0.1.0"
description = "Learn from PR reviews, help your team proofread faster and more efficiently"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyGithub>=2.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.3",
    "rich>=13.7.1"
]
